fix(rename):fix rename file and folder,now it can work correctly,when input are right{id:string or int,newname:string(no"/""\")}

+++ src/main/java/love/linyi/service/UserFolderService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserFolderService {
     int save(List<UserFolder> userFolderList);
     List<UserFolder> getUserFolderList(int id);
-    String reNameFileOrFolder(long id, String newName, int userId);
+    String reNameFileOrFolder(int id, String newName, int userId);
     UserFolder getUserFolderById(int id);
 
 }
